Fix #12356 Do not send keep-alive when not persistent
updates from review
