+++ jetty-core/jetty-server/src/test/java/org/eclipse/jetty/server/handler/AbstractConnectHandlerTest.java
@@ -11,14 +11,13 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.ee10.proxy;
+package org.eclipse.jetty.server.handler;
 
 import java.io.IOException;
 import java.net.Socket;
 
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.ServerConnector;
-import org.eclipse.jetty.server.handler.ConnectHandler;
 import org.junit.jupiter.api.AfterEach;
 
 public abstract class AbstractConnectHandlerTest
